fix: fix libraries versions conflict
 - add DI using Hilt
 - add API key to local properties
 - fix dto (Pexels API sends id of type Long, not Int)

+++ app/src/main/java/com/example/pexelsapp/data/models/PhotoDto.kt
@@ -5,13 +5,13 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class PhotoDto(
-    @SerialName("id") val id: Int,
+    @SerialName("id") val id: Long,
     @SerialName("width") val width: Int,
     @SerialName("height") val height: Int,
     @SerialName("url") val url: String,
     @SerialName("photographer") val photographer: String,
     @SerialName("photographer_url") val photographerUrl: String,
-    @SerialName("photographer_id") val photographerId: Int,
+    @SerialName("photographer_id") val photographerId: Long,
     @SerialName("avg_color") val avgColor: String,
     @SerialName("src") val src: PhotoSourcesDto,
     @SerialName("liked") val liked: Boolean,
